[rafactor] 대회 일정 리팩토링 진행
1. 컴포넌트 내에서 사용된 함수들을 따로 분리해서 저장
2. 대회가 없는 경우 대회가 없다고 알려주도록 수정
